+++ Assets/Scripts/Productions/ProductionBois.cs
@@ -10,20 +10,20 @@ namespace Production
         /// Production des bucherons avec le bonus des scieries.
         /// </summary>
         /// <returns>Production de bois</returns>
-        public override int CalculerProduction()
+        public override long CalculerProduction()
         {
             // Calcul de l'effort des bucherons
             int professionPourcent = GestionnaireProfessions.Instance.professionDict[ProfessionEnum.BUCHERON].professionPourcent;
-            int popActuelle = InventaireRessources.Instance.AccesQteRessource(RessourceEnum.POPULATION);
-            int nbPopTravaille = popActuelle * professionPourcent / 100;
+            long popActuelle = InventaireRessources.Instance.AccesQteRessource(RessourceEnum.POPULATION);
+            long nbPopTravaille = popActuelle * professionPourcent / 100;
             if (nbPopTravaille <= 0)
                 return 0;
             // Calcul du bonus de production des scieries
             float bonusScieries = (float)((ScierieConfig)GestionnaireBatiments.Instance.batimentConfigDict[BatimentEnum.SCIERIE]).bonusProductionPourcent / 100;
-            int nbFermes = GestionnaireBatiments.Instance.AccesQteBatiment(BatimentEnum.SCIERIE);
+            long nbFermes = GestionnaireBatiments.Instance.AccesQteBatiment(BatimentEnum.SCIERIE);
             bonusScieries = bonusScieries * nbFermes + 1;
 
-            return (int)(nbPopTravaille * efficacitePourcent / 100 * bonusScieries);
+            return (long)(nbPopTravaille * efficacitePourcent / 100 * bonusScieries);
         }
     }
 }
